UserDefaults에 저장된 사용자 값 유무에 따른 시작화면 처리, viewcontroller 파일 분리
