+++ MuziGit/Base.lproj/Main.storyboard
@@ -8,10 +8,10 @@
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
     <scenes>
-        <!--View Controller-->
+        <!--Onboarding View Controller-->
         <scene sceneID="tne-QT-ifu">
             <objects>
-                <viewController storyboardIdentifier="MainPage" id="BYZ-38-t0r" customClass="ViewController" customModule="MuziGit" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController storyboardIdentifier="Main" id="BYZ-38-t0r" customClass="OnboardingViewController" customModule="MuziGit" customModuleProvider="target" sceneMemberID="viewController">
                     <view key="view" contentMode="scaleToFill" id="8bC-Xf-vdC">
                         <rect key="frame" x="0.0" y="0.0" width="393" height="852"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
@@ -28,7 +28,7 @@
                                 <state key="normal" title="Button"/>
                                 <buttonConfiguration key="configuration" style="plain" title="Github로 로그인"/>
                                 <connections>
-                                    <action selector="touchUpGithubLogin:" destination="BYZ-38-t0r" eventType="touchUpInside" id="4bl-GM-SPg"/>
+                                    <action selector="touchUpGithubLogin:" destination="BYZ-38-t0r" eventType="touchUpInside" id="0YP-Us-ZKL"/>
                                 </connections>
                             </button>
                             <button opaque="NO" contentMode="scaleToFill" fixedFrame="YES" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="system" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="kle-tN-M7e">
@@ -37,7 +37,7 @@
                                 <state key="normal" title="Button"/>
                                 <buttonConfiguration key="configuration" style="plain" title="Button"/>
                                 <connections>
-                                    <action selector="check:" destination="BYZ-38-t0r" eventType="touchUpInside" id="0Bq-S4-gLt"/>
+                                    <action selector="check:" destination="BYZ-38-t0r" eventType="touchUpInside" id="w3O-dC-vPw"/>
                                 </connections>
                             </button>
                             <containerView opaque="NO" contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="nPi-7d-JnP">
@@ -46,6 +46,13 @@
                                     <segue destination="6nP-AB-hin" kind="embed" id="oWg-p2-M7c"/>
                                 </connections>
                             </containerView>
+                            <pageControl opaque="NO" contentMode="scaleToFill" fixedFrame="YES" contentHorizontalAlignment="center" contentVerticalAlignment="center" numberOfPages="3" translatesAutoresizingMaskIntoConstraints="NO" id="3sz-5e-O6g">
+                                <rect key="frame" x="138" y="661" width="116" height="26"/>
+                                <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
+                                <connections>
+                                    <action selector="changePageValue:" destination="BYZ-38-t0r" eventType="valueChanged" id="KIE-5T-0RW"/>
+                                </connections>
+                            </pageControl>
                         </subviews>
                         <viewLayoutGuide key="safeArea" id="6Tk-OE-BBY"/>
                         <color key="backgroundColor" systemColor="systemBackgroundColor"/>
@@ -59,7 +66,7 @@
                         </constraints>
                     </view>
                     <connections>
-                        <outlet property="onboardingPageControl" destination="lf7-sx-sIW" id="gbl-J7-B8p"/>
+                        <outlet property="onboardingPageControl" destination="3sz-5e-O6g" id="EkI-9O-9Lg"/>
                     </connections>
                 </viewController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="dkx-z0-nzr" sceneMemberID="firstResponder"/>
